feat: serde performance level
